+++ lib/rubocop/config_loader_resolver.rb
@@ -155,7 +155,7 @@ module RuboCop
 
     def gem_config_path(gem_name, relative_config_path)
       spec = Gem::Specification.find_by_name(gem_name)
-      return File.join(spec.gem_dir, relative_config_path)
+      File.join(spec.gem_dir, relative_config_path)
     rescue Gem::LoadError => e
       raise Gem::LoadError,
             "Unable to find gem #{gem_name}; is the gem installed? #{e}"